Add second hydro chart
